+++ src/commands/Article.ts
@@ -1,3 +1,13 @@
+import {
+  Position,
+  TextDocument,
+  TextDocumentWillSaveEvent,
+  TextEdit,
+  Uri,
+  commands,
+  window,
+  workspace
+} from 'vscode';
 import { Folders } from './Folders';
 import { DEFAULT_CONTENT_TYPE } from './../constants/ContentType';
 import { isValidFile } from './../helpers/isValidFile';
@@ -13,7 +23,6 @@ import {
   TelemetryEvent,
   SETTING_SLUG_TEMPLATE
 } from './../constants';
-import * as vscode from 'vscode';
 import { CustomPlaceholder, Field } from '../models';
 import { format } from 'date-fns';
 import {
@@ -33,17 +42,35 @@ import { Telemetry } from '../helpers/Telemetry';
 import { ParsedFrontMatter } from '../parsers';
 import { MediaListener } from '../listeners/panel';
 import { NavigationType } from '../dashboardWebView/models';
-import { Position } from 'vscode';
 import { SNIPPET } from '../constants/Snippet';
 import * as l10n from '@vscode/l10n';
 import { LocalizationKey } from '../localization';
 
 export class Article {
   /**
+   * Registers the commands for the Article class.
+   *
+   * @param subscriptions - The array of subscriptions to register the commands with.
+   */
+  public static async registerCommands(subscriptions: unknown[]) {
+    subscriptions.push(
+      commands.registerCommand(COMMAND_NAME.setLastModifiedDate, Article.setLastModifiedDate)
+    );
+
+    subscriptions.push(commands.registerCommand(COMMAND_NAME.generateSlug, Article.updateSlug));
+
+    // Inserting an image in Markdown
+    subscriptions.push(commands.registerCommand(COMMAND_NAME.insertMedia, Article.insertMedia));
+
+    // Inserting a snippet in Markdown
+    subscriptions.push(commands.registerCommand(COMMAND_NAME.insertSnippet, Article.insertSnippet));
+  }
+
+  /**
    * Sets the article date
    */
   public static async setDate() {
-    const editor = vscode.window.activeTextEditor;
+    const editor = window.activeTextEditor;
     if (!editor) {
       return;
     }
@@ -77,7 +104,7 @@ export class Article {
    * Sets the article lastmod date
    */
   public static async setLastModifiedDate() {
-    const editor = vscode.window.activeTextEditor;
+    const editor = window.activeTextEditor;
     if (!editor) {
       return;
     }
@@ -91,9 +118,7 @@ export class Article {
     ArticleHelper.update(editor, updatedArticle as ParsedFrontMatter);
   }
 
-  public static async setLastModifiedDateOnSave(
-    document: vscode.TextDocument
-  ): Promise<vscode.TextEdit[]> {
+  public static async setLastModifiedDateOnSave(document: TextDocument): Promise<TextEdit[]> {
     const updatedArticle = this.setLastModifiedDateInner(document);
 
     if (typeof updatedArticle === 'undefined') {
@@ -105,9 +130,7 @@ export class Article {
     return [update];
   }
 
-  private static setLastModifiedDateInner(
-    document: vscode.TextDocument
-  ): ParsedFrontMatter | undefined {
+  private static setLastModifiedDateInner(document: TextDocument): ParsedFrontMatter | undefined {
     const article = ArticleHelper.getFrontMatterFromDocument(document);
 
     // Only set the date, if there is already front matter set
@@ -160,7 +183,7 @@ export class Article {
     Telemetry.send(TelemetryEvent.generateSlug);
 
     const updateFileName = Settings.get(SETTING_SLUG_UPDATE_FILE_NAME) as string;
-    const editor = vscode.window.activeTextEditor;
+    const editor = window.activeTextEditor;
 
     if (!editor) {
       return;
@@ -219,7 +242,7 @@ export class Article {
       // Check if the file name should be updated by the slug
       // This is required for systems like Jekyll
       if (updateFileName) {
-        const editor = vscode.window.activeTextEditor;
+        const editor = window.activeTextEditor;
         if (editor) {
           const ext = extname(editor.document.fileName);
           const fileName = basename(editor.document.fileName);
@@ -237,7 +260,7 @@ export class Article {
           try {
             await editor.document.save();
 
-            await vscode.workspace.fs.rename(editor.document.uri, vscode.Uri.file(newPath), {
+            await workspace.fs.rename(editor.document.uri, Uri.file(newPath), {
               overwrite: false
             });
           } catch (e: unknown) {
@@ -257,7 +280,7 @@ export class Article {
    * Retrieve the slug from the front matter
    */
   public static getSlug() {
-    const editor = vscode.window.activeTextEditor;
+    const editor = window.activeTextEditor;
     if (!editor) {
       return;
     }
@@ -297,7 +320,7 @@ export class Article {
    * Toggle the page its draft mode
    */
   public static async toggleDraft() {
-    const editor = vscode.window.activeTextEditor;
+    const editor = window.activeTextEditor;
     if (!editor) {
       return;
     }
@@ -315,7 +338,7 @@ export class Article {
    * Article auto updater
    * @param event
    */
-  public static async autoUpdate(event: vscode.TextDocumentWillSaveEvent) {
+  public static async autoUpdate(event: TextDocumentWillSaveEvent) {
     const document = event.document;
     if (document && ArticleHelper.isSupportedFile(document)) {
       const autoUpdate = Settings.get(SETTING_AUTO_UPDATE_DATE);
@@ -355,7 +378,7 @@ export class Article {
    * Insert an image from the media dashboard into the article
    */
   public static async insertMedia() {
-    const editor = vscode.window.activeTextEditor;
+    const editor = window.activeTextEditor;
     if (!editor) {
       return;
     }
@@ -367,7 +390,7 @@ export class Article {
     const position = editor.selection.active;
     const selectionText = editor.document.getText(editor.selection);
 
-    await vscode.commands.executeCommand(COMMAND_NAME.dashboard, {
+    await commands.executeCommand(COMMAND_NAME.dashboard, {
       type: 'media',
       data: {
         pageBundle: !!contentType.pageBundle,
@@ -386,7 +409,7 @@ export class Article {
    * Insert a snippet into the article
    */
   public static async insertSnippet() {
-    const editor = vscode.window.activeTextEditor;
+    const editor = window.activeTextEditor;
     if (!editor) {
       return;
     }
@@ -442,7 +465,7 @@ export class Article {
     const article = ArticleHelper.getFrontMatter(editor);
     const contentType = article ? ArticleHelper.getContentType(article) : undefined;
 
-    await vscode.commands.executeCommand(COMMAND_NAME.dashboard, {
+    await commands.executeCommand(COMMAND_NAME.dashboard, {
       type: NavigationType.Snippets,
       data: {
         fileTitle: article?.data.title || '',